pyproject.toml
[tool.poetry]
name = "fakefill"
version = "1.0.0"
description = ""
authors = ["Ben CHEN <bn@benbenbang.io>"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3.1"
loguru = "^0.5.2"

[tool.poetry.dev-dependencies]
apache-airflow = "^1.10.12"
psycopg2-binary = "^2.8.6"
twine = "^3.4.1"
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
fakefill="fakefill.cli:cli"
